requirements.txt

DateTime==5.5
Flask==3.1.1
flask-cors==6.0.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
seaborn==0.13.2
seleniumbase==4.38.3
wordcloud==1.9.4
python-dotenv==1.1.0